add permissions validation to user and entry services

+++ src/controllers/entryController.js
@@ -2,8 +2,12 @@ import entryService from "../services/entryService.js";
 
 async function create(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { userId } = req.params;
-    const newEntry = await entryService.create({ userId });
+    const newEntry = await entryService.create({ authorId: userId, currentUser });
     res.status(201).json(newEntry);
   } catch (error) {
     next(error);
@@ -31,11 +35,15 @@ async function getById(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { entryId } = req.params;
     const { title } = req.body;
     // Not passing Blocks Array because blocks should be created with an entryId already, 
     // even when editing, no need to directly move a block to an entry
-    const updatedEntry = await entryService.update(entryId, { title });
+    const updatedEntry = await entryService.update({ id: entryId, title, currentUser });
     res.status(200).json(updatedEntry);
   } catch (error) {
     next(error);
@@ -43,8 +51,12 @@ async function update(req, res, next) {
 }
 async function deleteById(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { entryId } = req.params;
-    await entryService.deleteById(entryId);
+    await entryService.deleteById({ id: entryId, currentUser });
     res.status(204);
   } catch (error) {
     next(error);
