+++ src/controllers/userController.js
@@ -2,8 +2,12 @@ import userService from "../services/userService.js";
 
 async function getOne(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { userId } = req.params;
-    const user = await userService.getById(userId);
+    const user = await userService.getById({ id: userId, currentUser });
     res.status(200).json(user);
   } catch (error) {
     next(error);
@@ -12,7 +16,11 @@ async function getOne(req, res, next) {
 
 async function getAll(req, res, next) {
   try {
-    const users = await userService.getAll();
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
+    const users = await userService.getAll({ currentUser });
     res.status(200).json(users);
   } catch (error) {
     next(error);
@@ -30,9 +38,13 @@ async function create(req, res, next) {
 }
 async function update(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { userId } = req.params;
     const { name, email, password } = req.body;
-    const updatedUser = await userService.update(userId, { name, email, password });
+    const updatedUser = await userService.update({ id: userId, input: { name, email, password }, currentUser });
     res.status(200).json(updatedUser);
   } catch (error) {
     next(error);
@@ -42,7 +54,7 @@ async function updateRole(req, res, next) {
   try {
     const { userId } = req.params;
     const { role } = req.body;
-    const updatedUser = await userService.updateRole({ id: userId, role });
+    const updatedUser = await userService.updateRole({ id: userId, role, currentUser: req.user });
     res.status(200).json(updatedUser);
   } catch (error) {
     next(error);
@@ -51,7 +63,7 @@ async function updateRole(req, res, next) {
 async function deleteById(req, res, next) {
   try {
     const { userId } = req.params;
-    await userService.deleteById(userId);
+    await userService.deleteById({ id: userId, currentUser: req.user });
     res.status(200).json({
       'message': 'User deleted succesfully'
     });
